4-3.db/3_graph_falkor/requirements.txt
ifcopenshell
neo4j
python-dotenv
langchain
langchain-community
langchain-ollama
ollama
streamlit